[master] updated code

+++ src/common/lcg.c
@@ -38,5 +38,5 @@ static u32 lcg_parkmiller_state_internal(void) {
 }
 
 u64 lcg_parkmiller_64(void) {
-    return (u64)((lcg_parkmiller_state_internal() << 32) + lcg_parkmiller_state_internal());
+    return (u64)(((u64)(lcg_parkmiller_state_internal()) << 32) + lcg_parkmiller_state_internal());
 }
